+++ src/libmerc/tls.h
@@ -23,6 +23,10 @@
 #include "mem_utils.hpp"
 
 
+// Forward declaration for crypto_policy::required_extensions
+namespace crypto_policy {
+    struct required_extensions;
+}
 
 // class xtn represents a TLS extension
 //
@@ -120,10 +124,13 @@ enum class tls_content_type : uint16_t {
 };
 
 enum class tls_version : uint16_t {
+    none = 0x0000,
+    sslv2_0 = 0x0002,
     sslv3_0 = 0x0300,
     tlsv1_0 = 0x0301,
     tlsv1_1 = 0x0302,
     tlsv1_2 = 0x0303,
+    tlsv1_3 = 0x0304,
 };
 
 /*
@@ -294,6 +301,7 @@ struct tls_server_certificate {
     bool is_not_empty() const { return certificate_list.is_not_empty(); }
 
     void write_json(struct json_array &a, bool json_output) const;
+    bool get_subject_common_name(std::string &common_name) const;
 
     static constexpr mask_and_value<8> matcher{
         { 0xff, 0xff, 0xfc, 0x00, 0x00, 0xff, 0x00, 0x00 },
@@ -352,6 +360,10 @@ struct tls_extensions : public datum {
     void write_raw_features(writeable &buf) const;
 
     datum get_supported_groups() const;
+
+    tls_version get_supported_versions() const;
+
+    crypto_policy::required_extensions get_required_extensions() const;
 
     void write_l7_metadata(cbor_object &o) const {
         o.print_key_string("server_name", get_server_name());
@@ -536,6 +548,13 @@ struct tls_server_hello : public base_protocol {
         { 0x16, 0x03, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00 }
     };
 
+    tls_version get_version() const {
+        struct datum tmp = protocol_version;
+        uint16_t version;
+        tmp.read_uint16(&version);
+        return (tls_version)version;
+    }
+
 };
 
 class tls_server_hello_and_certificate : public base_protocol {
@@ -629,6 +648,10 @@ public:
     }
 
     const tls_server_hello & get_server_hello() const { return hello; }
+
+    bool get_subject_common_name(std::string &common_name) const {
+        return certificate.get_subject_common_name(common_name);
+    }
 };
 
 
@@ -727,6 +750,10 @@ public:
             tls.close();
 
         }
+    }
+
+    bool get_subject_common_name(std::string &common_name) const {
+        return certificate.get_subject_common_name(common_name);
     }
 
 };
@@ -951,6 +978,8 @@ inline bool is_faketls_util(const datum ciphersuite_vector) {
 #define type_session_ticket                  0x0023
 #define type_quic_transport_parameters       0x0039
 #define type_quic_transport_parameters_draft 0xffa5
+#define type_ec_points_format                0x000b
+#define type_encrypt_then_mac                0x0016
 
 #define type_ech_client_hello                0xfe0d
 
@@ -2152,6 +2181,31 @@ inline void tls_server_certificate::write_json(struct json_array &a, bool json_o
             return;
         }
     }
+}
+
+inline bool tls_server_certificate::get_subject_common_name(std::string &common_name) const {
+    struct datum tmp_cert_list = certificate_list;
+    while (tmp_cert_list.length() > 0) {
+        uint64_t tmp_len;
+        if (tmp_cert_list.read_uint(&tmp_len, L_CertificateLength) == false) {
+            return false;
+        }
+        if (tmp_len > (unsigned)tmp_cert_list.length()) {
+            tmp_len = tmp_cert_list.length(); /* truncate */
+        }
+        if (tmp_len == 0) {
+            return false;
+        }
+        struct x509_cert c;
+        c.parse(tmp_cert_list.data, tmp_len);
+        if (c.get_subject_common_name(common_name)) {
+            return true;
+        }
+        if (tmp_cert_list.skip(tmp_len) == false) {
+            return false;
+        }
+    }
+    return false;
 }