+++ doc/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Enabled IPv6 longest prefix matching for:
   * ASNs mapped to IPv6 subnets.
   * Domain names mapped to IPv6 subnets for Domain faking detection.
+* Added NIST SP800-52-2 cryptographic assessment.
 
 ## VERSION 2.13.1
 * Fixed potential underflows and out-of-bounds accesses in TCP and QUIC